+++ harness_sdk/executor_plugins/__init__.py
@@ -5,6 +5,9 @@ from __future__ import annotations
 from harness_sdk.executor_plugins.base import ExecutorPluginError, VerifyExecutor
 from harness_sdk.executor_plugins.dry_run import DryRunExecutor
 from harness_sdk.executor_plugins.preview import PreviewExecutor
+from harness_sdk.executor_plugins.sandbox import SandboxConfigError, SandboxExecutor
+from harness_sdk.executor_plugins.docker import DockerExecutor
+from harness_sdk.executor_plugins.firejail import FirejailExecutor
 from harness_sdk.executor_plugins.subprocess import SubprocessExecutor
 from harness_sdk.executor_plugins._loader import load_executor_plugin
 
@@ -12,6 +15,10 @@ __all__ = [
     "DryRunExecutor",
     "ExecutorPluginError",
     "PreviewExecutor",
+    "SandboxConfigError",
+    "SandboxExecutor",
+    "DockerExecutor",
+    "FirejailExecutor",
     "SubprocessExecutor",
     "VerifyExecutor",
     "load_executor_plugin",